add psat report rollup service to seperate psat logic
