+++ src/app/report-rollup/psat-report-rollup.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { PsatReportRollupService } from './psat-report-rollup.service';
+
+describe('PsatReportRollupService', () => {
+  let service: PsatReportRollupService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(PsatReportRollupService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
