+++ src/app/report-rollup/report-summary/psat-summary/psat-summary.component.html
@@ -2,7 +2,7 @@
     <div class="card-header">
         <h3>Pumps</h3>
         <span>
-            <span>{{reportRollupService.numPsats}}</span>
+            <span>{{numPsats}}</span>
         </span>
     </div>
     <div class="card-body">
